+++ grails-app/conf/FileUploaderBootStrap.groovy
@@ -1,22 +1,8 @@
-import com.causecode.fileuploader.RackspaceCDNFileUploaderService
-
 class FileUploaderBootStrap {
 
-    def fileUploaderService
-
-    RackspaceCDNFileUploaderService rackspaceCDNFileUploaderService
-
     def init = { servletContext ->
-        log.debug "Fileuploader bootstap started executing."
-        if(rackspaceCDNFileUploaderService.authenticate()) {
-            rackspaceCDNFileUploaderService.listContainers()
-        }
-
-        log.debug "Fileuploader bootstap finished executing."
     }
 
     def destroy = {
-        rackspaceCDNFileUploaderService.close()
-        log.info "Fileuploader Bootstrap destroyed."
     }
 }